Reorg auth
