--- homeassistant/auth_providers/__init__.py
@@ -1,1 +1,0 @@
-"""Auth providers for Home Assistant."""
